requirements.txt
reflex>=0.5.10
litellm~=1.65.0

pydantic~=2.11.1
python-dotenv